+++ internal/vba/procedureir/clone.go
@@ -94,6 +94,7 @@ func cloneProcedure(in ProcedureIR) ProcedureIR {
 func cloneParameters(in []Parameter) []Parameter {
 	out := append([]Parameter(nil), in...)
 	for i := range out {
+		out[i].PassingRange = cloneRangePointer(in[i].PassingRange)
 		out[i].DefaultRange = cloneRangePointer(in[i].DefaultRange)
 		out[i].BoundsRange = cloneRangePointer(in[i].BoundsRange)
 		out[i].ArrayBounds = append([]ArrayBound(nil), in[i].ArrayBounds...)